django/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.2
cffi==2.0.0
Django==5.2.6
gitdb==4.0.12
GitPython==3.1.41
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.23
pystache==0.6.8
setuptools==75.6.0
smmap==5.0.2
sqlparse==0.5.3
